scripts: check_init_priorities: add ltdc driver to ignore list
STM32 LTDC driver requires reference to display set
in DTS to pass calls.
When using MIPI DSI display it is not possible to
do it, as LTDC has to be already initialized for DSI host
to work, but initialization order check prevents
from referencing devices that will be initialized later.

This change should be removed when proper solution is
added to mainline.