modules: lvgl_display: use DMA2D to transfer buffer to fb
Use DMA2D to transfer LVGL buffer to LTDC framebuffer.
It allows to offload data transfer and free CPU, as function can
return and interrupt is used to determine if transfer was completed.
It also offers significant performance gains due to direct and
fast access to both internal and external RAM memories.